+++ packages/@glimmer/destroyable/package.json
@@ -1,0 +1,13 @@
+{
+  "name": "@glimmer/destroyable",
+  "version": "0.67.0",
+  "repository": "https://github.com/glimmerjs/glimmer-vm/tree/master/packages/@glimmer/destroyable",
+  "description": "Utilities for creating and managing a destroyable hierarchy of objects",
+  "license": "MIT",
+  "dependencies": {
+    "@glimmer/env": "0.1.7",
+    "@glimmer/global-context": "0.67.0",
+    "@glimmer/interfaces": "0.67.0",
+    "@glimmer/util": "0.67.0"
+  }
+}
