+++ packages/@glimmer/owner/package.json
@@ -1,0 +1,10 @@
+{
+  "name": "@glimmer/owner",
+  "version": "0.67.0",
+  "repository": "https://github.com/glimmerjs/glimmer-vm/tree/master/packages/@glimmer/owner",
+  "description": "Implementation for the owner in Glimmer apps",
+  "license": "MIT",
+  "dependencies": {
+    "@glimmer/util": "0.67.0"
+  }
+}
