+++ packages/@glimmer/runtime/package.json
@@ -13,6 +13,8 @@
     "@glimmer/validator": "0.67.0",
     "@glimmer/wire-format": "0.67.0",
     "@glimmer/interfaces": "0.67.0",
+    "@glimmer/destroyable": "0.67.0",
+    "@glimmer/owner": "0.67.0",
     "@glimmer/program": "0.67.0",
     "@glimmer/vm": "0.67.0",
     "@simple-dom/interface": "^1.4.0"
